fix(docs): keep pagefind search mounted across doc navigation
Search stopped working after any client-side docs navigation because
DocsShell rendered the pagefind-modal/pagefind-modal-trigger pair from
[[...slug]]/page.tsx, which the App Router fully remounts on every
dynamic-segment change. Pagefind pairs the trigger and modal via
generated IDs the first time each connects, and that pairing doesn't
survive being torn down and recreated -- only a hard refresh (a fresh
document) reset it.

Split the docs page into a persistent layout.tsx + DocsShell (sidebar,
search modal/trigger, mobile compact bar -- mounts once per docs
session) and a per-doc DocsArticle (article + TOC -- remounts on every
doc, as intended). Also drops the lastSidebarScrollTop module-level
hack, no longer needed now that the sidebar is a real persistent DOM
node.
